Now enforcing the use of query-string style parameters on the <script> node.
Setting the path to dependant files is now done in at the end of the "src"
attribute for the HtmlMedia.js file, query-string style! i.e. HtmlMedia.js?path=someFolder

Also, now there's an event that the developer can hook into in order to
be confident that <audio> and <video> nodes on the page are properly treated
before interacting with them. It's called "onready" and currently you can
define the name of a global function to call in the query-string src:
i.e. HtmlMedia.js?path=folder&onready=myinit
OR include a "onready" attribute on the <script> node. The content of the
attribute needs to be raw JavaScript, not the name of a function.